Comments (for vault only at the moment, but the structure is there for news/journals/etc.)

+++ database/migrations/2015_04_13_102856_create_vault_items_table.php
@@ -28,7 +28,7 @@ class CreateVaultItemsTable extends Migration {
             $table->integer('stat_downloads');
             $table->integer('stat_ratings');
             $table->integer('stat_comments');
-            $table->integer('stat_average_rating');
+            $table->decimal('stat_average_rating');
 			$table->timestamps();
             $table->softDeletes();
